Apply Eric's comments
