+++ docker/rucio_client/scripts/k8s_sync_sites.sh
@@ -40,8 +40,8 @@ echo "Set site capacity"
 ./setSiteCapacity
 
 echo "Setting availability"
-./setSiteAvailability.py
+./setSiteAvailability
 echo "Setting wmcore_output_tape attribute"
-./set_wmcore_output_tape.py
+./setWmcoreTapeOutput
 echo "Update DDM quota"
 ./updateDDMQuota
